+++ apps/packaged/src/errors.ts
@@ -1,0 +1,9 @@
+export class PackagedPathAccessError extends Error {
+  readonly title: string;
+
+  constructor(message: string, options?: { cause?: unknown; title?: string }) {
+    super(message, options);
+    this.name = "PackagedPathAccessError";
+    this.title = options?.title ?? "Open Design cannot access its data folder";
+  }
+}
