+++ apps/packaged/src/launch.ts
@@ -5,24 +5,14 @@ import { userInfo } from "node:os";
 
 import { app } from "electron";
 
+import { PackagedPathAccessError } from "./errors.js";
 import type { PackagedNamespacePaths } from "./paths.js";
-
-export class PackagedPathAccessError extends Error {
-  readonly title: string;
-
-  constructor(message: string, options?: { cause?: unknown; title?: string }) {
-    super(message, options);
-    this.name = "PackagedPathAccessError";
-    this.title = options?.title ?? "Open Design cannot access its data folder";
-  }
-}
 
 export type PackagedSingleInstanceApp = {
   on: (event: "second-instance", listener: () => void) => unknown;
   quit: () => void;
   requestSingleInstanceLock: () => boolean;
 };
-
 type PathDiagnostic = {
   exists: boolean;
   mode?: number;
